Various warning fixes (#663)
* various warning fixes

* extra log statement for invalid ucode

* Update gfx_pc.cpp

* revert config change

+++ src/config/Config.cpp
@@ -137,7 +137,7 @@ void Config::EraseBlock(const std::string& key) {
         nlohmann::json& gjson2 = gjson;
         std::vector<std::string> dots = StringHelper::Split(key, ".");
         if (dots.size() > 1) {
-            int curDot = 0;
+            size_t curDot = 0;
             for (auto& dot : dots) {
                 if (gjson2.contains(dot)) {
                     if (curDot == dots.size()) {
